+++ dsmrreader/config/base.py
@@ -198,5 +198,9 @@ DSMRREADER_LATEST_VERSION_FILE = 'https://raw.githubusercontent.com/dennissiemen
 
 DSMRREADER_REST_FRAMEWORK_API_USER = 'api-user'
 
+# Sleep durations for infinity processes. Update these in your own config if you wish to alter them.
+DSMRREADER_BACKEND_SLEEP = 1
+DSMRREADER_DATALOGGER_SLEEP = 0.25
+
 # Whether telegrams are logged, in base64 format. Only required for debugging.
 DSMRREADER_LOG_TELEGRAMS = False