+++ docs/changelog.rst
@@ -18,7 +18,7 @@ Please make sure you have a fresh **database backup** before upgrading! Upgradin
 
 
 
-v1.12.0 - 2018-xx-xx
+v1.12.0 - 2018-01-14
 ^^^^^^^^^^^^^^^^^^^^
 
 **Tickets resolved in this release:**